feature: moved editing screening params to a separate file
