+++ src/bot/core/bot.py
@@ -8,6 +8,7 @@ from src.bot.config import TELEGRAM_BOT_TOKEN
 
 # Основные роутеры
 from src.bot.handlers.admin_commands import admin_router
+from src.bot.handlers.admin_screening_commands import admin_screening
 from src.bot.handlers.hr_registration import hr_registration_router
 from src.bot.handlers.hr_commands import hr_commands_router
 from src.bot.handlers.candidate_commands import candidate_router
@@ -27,8 +28,9 @@ dp = Dispatcher(storage=memory)
 # Внимание! 
 # Порядок регистрации роутеров имеет значение
 dp.include_routers(
-    admin_router,
     tests_router,
+    admin_router,
+    admin_screening,
     candidate_router,
     hr_registration_router,
     hr_commands_router,
